Extract Yaml from Content
